+++ BUILD
@@ -196,6 +196,7 @@ GRPC_PUBLIC_HDRS = [
     "include/grpc/byte_buffer.h",
     "include/grpc/byte_buffer_reader.h",
     "include/grpc/channel_credentials/google_default.h",
+    "include/grpc/channel_credentials/ssl.h",
     "include/grpc/compression.h",
     "include/grpc/fork.h",
     "include/grpc/grpc.h",
