+++ grpc.def
@@ -11,6 +11,16 @@ EXPORTS
     grpc_byte_buffer_reader_readall
     grpc_raw_byte_buffer_from_reader
     grpc_google_default_credentials_create
+    grpc_ssl_credentials_create
+    grpc_ssl_credentials_create_ex
+    grpc_ssl_server_certificate_config_create
+    grpc_ssl_server_certificate_config_destroy
+    grpc_ssl_server_credentials_create
+    grpc_ssl_server_credentials_create_ex
+    grpc_ssl_server_credentials_create_options_using_config
+    grpc_ssl_server_credentials_create_options_using_config_fetcher
+    grpc_ssl_server_credentials_options_destroy
+    grpc_ssl_server_credentials_create_with_options
     grpc_compression_algorithm_is_message
     grpc_compression_algorithm_is_stream
     grpc_compression_algorithm_parse
@@ -117,8 +127,6 @@ EXPORTS
     grpc_ssl_session_cache_create_channel_arg
     grpc_call_credentials_release
     grpc_set_ssl_roots_override_callback
-    grpc_ssl_credentials_create
-    grpc_ssl_credentials_create_ex
     grpc_composite_channel_credentials_create
     grpc_composite_call_credentials_create
     grpc_google_compute_engine_credentials_create
@@ -132,14 +140,6 @@ EXPORTS
     grpc_auth_metadata_context_copy
     grpc_auth_metadata_context_reset
     grpc_metadata_credentials_create_from_plugin
-    grpc_ssl_server_certificate_config_create
-    grpc_ssl_server_certificate_config_destroy
-    grpc_ssl_server_credentials_create
-    grpc_ssl_server_credentials_create_ex
-    grpc_ssl_server_credentials_create_options_using_config
-    grpc_ssl_server_credentials_create_options_using_config_fetcher
-    grpc_ssl_server_credentials_options_destroy
-    grpc_ssl_server_credentials_create_with_options
     grpc_call_set_credentials
     grpc_server_credentials_set_auth_metadata_processor
     grpc_alts_credentials_client_options_create
