+++ include/grpc/module.modulemap
@@ -6,6 +6,7 @@ header "byte_buffer.h"
   header "byte_buffer_reader.h"
   header "census.h"
   header "channel_credentials/google_default.h"
+  header "channel_credentials/ssl.h"
   header "compression.h"
   header "fork.h"
   header "grpc.h"
