+++ src/core/lib/security/credentials/google_default/google_default_credentials.cc
@@ -33,6 +33,7 @@
 #include "absl/types/optional.h"
 
 #include <grpc/channel_credentials/google_default.h>
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc_security.h>  // IWYU pragma: keep
 #include <grpc/grpc_security_constants.h>
 #include <grpc/impl/codegen/grpc_types.h>
