+++ src/core/lib/security/credentials/ssl/ssl_credentials.h
@@ -22,6 +22,7 @@
 
 #include <stddef.h>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
