+++ src/core/lib/security/security_connector/ssl/ssl_security_connector.h
@@ -22,6 +22,7 @@
 
 #include <stddef.h>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
