+++ src/cpp/client/secure_credentials.cc
@@ -30,6 +30,7 @@
 #include "absl/types/optional.h"
 
 #include <grpc/channel_credentials/google_default.h>
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/event_engine/event_engine.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/impl/codegen/gpr_types.h>
