+++ src/cpp/server/secure_server_credentials.cc
@@ -24,6 +24,7 @@
 #include <utility>
 #include <vector>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/slice.h>
 #include <grpc/status.h>
