+++ src/php/ext/grpc/channel_credentials.c
@@ -28,6 +28,7 @@
 #include <zend_exceptions.h>
 
 #include <grpc/channel_credentials/google_default.h>
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/string_util.h>
