+++ src/ruby/ext/grpc/rb_channel_credentials.c
@@ -26,6 +26,7 @@
 #include "rb_grpc.h"
 #include "rb_grpc_imports.generated.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/support/alloc.h>
