+++ src/ruby/ext/grpc/rb_server_credentials.c
@@ -23,6 +23,7 @@
 #include "rb_grpc.h"
 #include "rb_grpc_imports.generated.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/support/log.h>
