+++ test/core/bad_ssl/bad_ssl_test.cc
@@ -22,6 +22,7 @@
 
 #include <string>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/impl/codegen/propagation_bits.h>
