+++ test/core/bad_ssl/servers/alpn.cc
@@ -20,6 +20,7 @@
 
 #include <string.h>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
