+++ test/core/bad_ssl/servers/cert.cc
@@ -16,6 +16,7 @@
  *
  */
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
