+++ test/core/end2end/fixtures/h2_oauth2_tls13.cc
@@ -21,6 +21,7 @@
 
 #include <string>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
