+++ test/core/end2end/fixtures/h2_ssl_proxy.cc
@@ -18,6 +18,7 @@
 
 #include <string.h>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
