+++ test/core/end2end/fixtures/h2_ssl_tls12.cc
@@ -20,6 +20,7 @@
 
 #include <string>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
