+++ test/core/end2end/h2_ssl_session_reuse_test.cc
@@ -23,6 +23,7 @@
 
 #include <gtest/gtest.h>
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
