+++ test/core/handshake/server_ssl_common.cc
@@ -36,6 +36,7 @@
 #include "absl/base/thread_annotations.h"
 #include "absl/strings/str_cat.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
