+++ test/core/handshake/verify_peer_options.cc
@@ -34,6 +34,7 @@
 
 #include "absl/strings/str_cat.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/support/alloc.h>
