+++ test/core/surface/num_external_connectivity_watchers_test.cc
@@ -22,6 +22,7 @@
 
 #include "gtest/gtest.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
