+++ test/core/surface/public_headers_must_be_c89.c
@@ -20,6 +20,7 @@
 #include <grpc/byte_buffer_reader.h>
 #include <grpc/census.h>
 #include <grpc/channel_credentials/google_default.h>
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/compression.h>
 #include <grpc/fork.h>
 #include <grpc/grpc.h>
@@ -82,6 +83,16 @@ int main(int argc, char **argv) {
   printf("%lx", (unsigned long) grpc_byte_buffer_reader_readall);
   printf("%lx", (unsigned long) grpc_raw_byte_buffer_from_reader);
   printf("%lx", (unsigned long) grpc_google_default_credentials_create);
+  printf("%lx", (unsigned long) grpc_ssl_credentials_create);
+  printf("%lx", (unsigned long) grpc_ssl_credentials_create_ex);
+  printf("%lx", (unsigned long) grpc_ssl_server_certificate_config_create);
+  printf("%lx", (unsigned long) grpc_ssl_server_certificate_config_destroy);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_ex);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_options_using_config);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_options_using_config_fetcher);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_options_destroy);
+  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_with_options);
   printf("%lx", (unsigned long) grpc_compression_algorithm_is_message);
   printf("%lx", (unsigned long) grpc_compression_algorithm_is_stream);
   printf("%lx", (unsigned long) grpc_compression_algorithm_parse);
@@ -186,8 +197,6 @@ int main(int argc, char **argv) {
   printf("%lx", (unsigned long) grpc_ssl_session_cache_create_channel_arg);
   printf("%lx", (unsigned long) grpc_call_credentials_release);
   printf("%lx", (unsigned long) grpc_set_ssl_roots_override_callback);
-  printf("%lx", (unsigned long) grpc_ssl_credentials_create);
-  printf("%lx", (unsigned long) grpc_ssl_credentials_create_ex);
   printf("%lx", (unsigned long) grpc_composite_channel_credentials_create);
   printf("%lx", (unsigned long) grpc_composite_call_credentials_create);
   printf("%lx", (unsigned long) grpc_google_compute_engine_credentials_create);
@@ -201,14 +210,6 @@ int main(int argc, char **argv) {
   printf("%lx", (unsigned long) grpc_auth_metadata_context_copy);
   printf("%lx", (unsigned long) grpc_auth_metadata_context_reset);
   printf("%lx", (unsigned long) grpc_metadata_credentials_create_from_plugin);
-  printf("%lx", (unsigned long) grpc_ssl_server_certificate_config_create);
-  printf("%lx", (unsigned long) grpc_ssl_server_certificate_config_destroy);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_ex);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_options_using_config);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_options_using_config_fetcher);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_options_destroy);
-  printf("%lx", (unsigned long) grpc_ssl_server_credentials_create_with_options);
   printf("%lx", (unsigned long) grpc_call_set_credentials);
   printf("%lx", (unsigned long) grpc_server_credentials_set_auth_metadata_processor);
   printf("%lx", (unsigned long) grpc_alts_credentials_client_options_create);
