+++ test/core/surface/sequential_connectivity_test.cc
@@ -24,6 +24,7 @@
 
 #include "gtest/gtest.h"
 
+#include <grpc/channel_credentials/ssl.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/slice.h>
