+++ tools/doxygen/Doxyfile.core.internal
@@ -811,6 +811,7 @@ include/grpc/byte_buffer.h \
 include/grpc/byte_buffer_reader.h \
 include/grpc/census.h \
 include/grpc/channel_credentials/google_default.h \
+include/grpc/channel_credentials/ssl.h \
 include/grpc/compression.h \
 include/grpc/event_engine/endpoint_config.h \
 include/grpc/event_engine/event_engine.h \
